fix(admin): レビュー指摘事項を修正
- POST /api/admin/categories に P2002 ハンドリングを追加（slug重複時に409を返す）
- PUT /api/admin/categories/[id] に P2002 ハンドリングを追加（同上）
- inventoryUpdateSchema を category.ts から product.ts に移動（ドメイン整合）
- src/app/api/admin/inventory/route.ts のimportを product.ts に更新
