+++ src/app/api/admin/categories/route.ts
@@ -2,6 +2,7 @@
 import { NextRequest, NextResponse } from 'next/server'
 import { requireAdmin } from '@/lib/admin-auth'
 import { findAllCategoriesWithCount, createCategory } from '@/lib/db/categories'
+import { PrismaClientKnownRequestError } from '@/generated/prisma/internal/prismaNamespace'
 import { categorySchema } from '@/lib/validators/category'
 
 // GET /api/admin/categories - カテゴリ一覧取得
@@ -43,6 +44,13 @@ export const POST = async (req: NextRequest) => {
     const category = await createCategory(parsed.data)
     return NextResponse.json({ category }, { status: 201 })
   } catch (err) {
+    // 修正: slug ユニーク制約違反（P2002）を 409 で返す
+    if (err instanceof PrismaClientKnownRequestError && err.code === 'P2002') {
+      return NextResponse.json(
+        { error: 'このスラッグは既に使用されています', code: 'CONFLICT' },
+        { status: 409 }
+      )
+    }
     console.error('[admin/categories POST] エラー:', err)
     return NextResponse.json(
       { error: 'カテゴリの作成に失敗しました', code: 'INTERNAL_SERVER_ERROR' },
