+++ src/lib/validators/product.ts
@@ -17,3 +17,18 @@ export const productSchema = z.object({
 })
 
 export type ProductFormValues = z.infer<typeof productSchema>
+
+// 在庫一括更新スキーマ（管理画面用）
+export const inventoryUpdateSchema = z.object({
+  items: z.array(
+    z.object({
+      id: z.string().min(1, '商品IDは必須です'),
+      stock: z
+        .number()
+        .int('在庫数は整数で入力してください')
+        .min(0, '在庫数は0以上で入力してください'),
+    })
+  ),
+})
+
+export type InventoryUpdateValues = z.infer<typeof inventoryUpdateSchema>
